- Use the %ju conversion specifier for time_t and get rid of these non-portable (unsigned long) casts; replace some uint64_t with uintmax_t
git-svn-id: svn://svn.ircd-hybrid.org/svnroot/ircd-hybrid/trunk@6782 82007160-df01-0410-b94d-b575c5fd34c7

+++ modules/core/m_sjoin.c
@@ -156,8 +156,8 @@ ms_sjoin(struct Client *source_p, int parc, char *parv[])
     if (newts < 800000000)
     {
       sendto_realops_flags(UMODE_DEBUG, L_ALL, SEND_NOTICE,
-                           "*** Bogus TS %lu on %s ignored from %s(via %s)",
-                           (unsigned long)newts, chptr->name,
+                           "*** Bogus TS %ju on %s ignored from %s(via %s)",
+                           newts, chptr->name,
                            source_p->name, source_p->from->name);
 
       newts = (oldts == 0) ? 0 : 800000000;
@@ -168,11 +168,11 @@ ms_sjoin(struct Client *source_p, int parc, char *parv[])
     if (!newts && !isnew && oldts)
     {
       sendto_channel_local(NULL, chptr, 0, 0, 0,
-                           ":%s NOTICE %s :*** Notice -- TS for %s changed from %lu to 0",
-                           me.name, chptr->name, chptr->name, (unsigned long)oldts);
+                           ":%s NOTICE %s :*** Notice -- TS for %s changed from %ju to 0",
+                           me.name, chptr->name, chptr->name, oldts);
       sendto_realops_flags(UMODE_SERVNOTICE, L_ALL, SEND_NOTICE,
-                           "Server %s changing TS on %s from %lu to 0",
-                           source_p->name, chptr->name, (unsigned long)oldts);
+                           "Server %s changing TS on %s from %ju to 0",
+                           source_p->name, chptr->name, oldts);
     }
   }
 
@@ -239,9 +239,9 @@ ms_sjoin(struct Client *source_p, int parc, char *parv[])
     }
 
     sendto_channel_local(NULL, chptr, 0, 0, 0,
-                         ":%s NOTICE %s :*** Notice -- TS for %s changed from %lu to %lu",
+                         ":%s NOTICE %s :*** Notice -- TS for %s changed from %ju to %ju",
                          me.name, chptr->name, chptr->name,
-                         (unsigned long)oldts, (unsigned long)newts);
+                         oldts, newts);
   }
 
   if (*modebuf != '\0')
@@ -260,8 +260,8 @@ ms_sjoin(struct Client *source_p, int parc, char *parv[])
     modebuf[1] = '\0';
   }
 
-  buflen = snprintf(uid_buf, sizeof(uid_buf), ":%s SJOIN %lu %s %s %s:",
-                    source_p->id, (unsigned long)tstosend,
+  buflen = snprintf(uid_buf, sizeof(uid_buf), ":%s SJOIN %ju %s %s %s:",
+                    source_p->id, tstosend,
                     chptr->name, modebuf, parabuf);
   uid_ptr = uid_buf + buflen;
 
@@ -364,8 +364,8 @@ ms_sjoin(struct Client *source_p, int parc, char *parv[])
     {
       sendto_server(source_p, 0, 0, "%s", uid_buf);
 
-      buflen = snprintf(uid_buf, sizeof(uid_buf), ":%s SJOIN %lu %s %s %s:",
-                        source_p->id, (unsigned long)tstosend,
+      buflen = snprintf(uid_buf, sizeof(uid_buf), ":%s SJOIN %ju %s %s %s:",
+                        source_p->id, tstosend,
                         chptr->name, modebuf, parabuf);
       uid_ptr = uid_buf + buflen;
     }
